+++ test/unit/specs/lcdClient.spec.js
@@ -183,5 +183,10 @@ describe(`LCD Client`, () => {
     axios.get = jest.fn().mockReturnValue({})
     client.getCandidate(`abc`)
     expect(axios.get.mock.calls).toMatchSnapshot()
-  })
+  }),
+    it(`queries a validator signing information`, async () => {
+      axios.get = jest.fn().mockReturnValue({})
+      client.queryValidatorSigningInfo(`pubKey`)
+      expect(axios.get.mock.calls).toMatchSnapshot()
+    })
 })