+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * possibility to end unbonding transactions @faboweb
 * close buttons to Staking modal @NodeGuy
 * more tests for new staking modal @NodeGuy
+* Add commission and uptime to LiValidator @fedekunze
+* Delete old bonding page @fedekunze
 
 ### Changed
 
@@ -26,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * cleaning up new validator profile page and balance header @jbibla
 * Changed a bunch of JavaScript files to strict mode. @NodeGuy @faboweb
 * prefer backquotes in code @ƒaboweb
+* Refactord submitDelegation. @NodeGuy
 
 ### Fixed